requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7  # Добавляем явное указание версии Werkzeug
pytest==7.0.1
black==21.12b0
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-pie==0.15.0
isort==5.9.3
mypy==0.931
